requirements.txt
youtube-transcript-api==1.2.3
google-api-python-client==2.149.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
PyGithub==2.5.0
